challenge2/requirements.txt
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dotenv==1.0.1
Scrapy==2.11.2
